backend/requirements.txt
pydantic-settings
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
python-multipart
python-jose[cryptography]
passlib[bcrypt]
easyocr
sentence-transformers
faiss-cpu
python-docx
openpyxl
pypdf
python-dotenv
pandas
numpy
pytest
httpx
firebase-admin
bcrypt==3.2.2
